chore(ui): factor git log to separate class
